requirements.txt
aiofiles==24.1.0
anyio==4.9.0
APScheduler==3.10.4
cachetools==4.2.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.3
decorator==5.2.1
future==1.0.0
git-filter-repo==2.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
loguru==0.7.2
pycparser==2.22
python-dotenv==1.0.0
python-telegram-bot==21.5
pytz==2025.2
requests==2.31.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
tornado==6.5.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
win32_setctime==1.2.0
SQLAlchemy